[Halpy-329] Move Create Case to Caseutils
